refactor: 테스트에 사용하는 Fixture 패키지 이동
